added the trigger after sending list and did a reformatting. Thinking on best way to use enums and channels to expand for multi channel devices
