+++ NV_ABJ/__init__.py
@@ -1,7 +1,16 @@
 ###################################################################################################################
 # Abstract Class Types
 ###################################################################################################################
-from NV_ABJ.abstract_class_types.ConnectedDevice import ConnectedDevice
+from NV_ABJ.abstract_device_interfaces.connected_device import ConnectedDevice
+from NV_ABJ.abstract_device_interfaces.signal_generator_interface import SignalGeneratorInterface, SignalGeneratorHardwareFormat
+from NV_ABJ.abstract_device_interfaces.photon_counter import PhotonCounter
+from NV_ABJ.abstract_device_interfaces.long_distance_positioner import LongDistancePositioner
+
+###################################################################################################################
+# User interfaces
+###################################################################################################################
+from NV_ABJ.user_interfaces.hardware_user_interfaces.signal_generator_gui.signal_generator_gui import SignalGeneratorGui
+from NV_ABJ.user_interfaces.hardware_user_interfaces.photon_counter_gui.photon_counter_gui import PhotonCounterGui
 
 
 ###################################################################################################################
@@ -11,19 +20,12 @@ from NV_ABJ.abstract_class_types.ConnectedDevice import ConnectedDevice
 #//////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 # Signal Generator
 #//////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-from NV_ABJ.hardware_interfaces.signal_generator.signal_generator_interface import SignalGeneratorInterface, SignalGeneratorHardwareFormat
-
-from NV_ABJ.user_interfaces.hardware_user_interfaces.signal_generator_gui.signal_generator_gui import SignalGeneratorGui
-# Hardware Imports ////////////////////////////////////////////////////////////////////////////////////////////////
-from NV_ABJ.hardware_interfaces.signal_generator.hardware.sg380.sg380 import SG380, ConfigurationSG380
+from NV_ABJ.hardware_interfaces.signal_generator.sg380.sg380 import SG380, ConfigurationSG380, ChannelsSG380
 
 #//////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 # Photon Counter
 #//////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-from NV_ABJ.hardware_interfaces.photon_counter.photon_counter import PhotonCounter
-from NV_ABJ.user_interfaces.hardware_user_interfaces.photon_counter_gui.photon_counter_gui import PhotonCounterGui
-# Hardware Imports ////////////////////////////////////////////////////////////////////////////////////////////////
-from NV_ABJ.hardware_interfaces.photon_counter.hardware.national_instruments_daq_counters.national_instruments_photon_counter_daq_controlled import NationalInstrumentsPhotonCounterDaqControlled
+from NV_ABJ.hardware_interfaces.photon_counter.national_instruments_daq_counters.national_instruments_photon_counter_daq_controlled import NationalInstrumentsPhotonCounterDaqControlled
 
 ###################################################################################################################
 # Logical Interfaces
