+++ NV_ABJ/abstract_device_interfaces/long_distance_positioner.py
@@ -1,0 +1,6 @@
+from abc import ABC, abstractmethod
+
+class LongDistancePositioner(ABC):
+
+    def __init__(self):
+        ...
