+++ NV_ABJ/abstract_device_interfaces/scanner.py
@@ -1,0 +1,7 @@
+from abc import ABCMeta, ABC,abstractmethod
+
+class SingleAxisScanner(metaclass=ABCMeta):
+    ...
+
+class XYZScanner(metaclass=ABCMeta):
+    ...
