+++ NV_ABJ/hardware_interfaces/photon_counter/national_instruments_daq_counters/__pycache__/national_instruments_photon_counter_daq_controlled.cpython-312.pyc
